requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
async-timeout==4.0.3
    # via redis
click==8.1.7
    # via flask
dnspython==2.6.1
    # via eventlet
eventlet==0.33.3
    # via -r requirements.in
flask==2.2.3
    # via
    #   -r requirements.in
    #   flask-cors
    #   flask-login
    #   flask-sqlalchemy
    #   flask-wtf
flask-cors==3.0.10
    # via -r requirements.in
flask-login==0.6.1
    # via -r requirements.in
flask-sqlalchemy==3.0.3
    # via -r requirements.in
flask-wtf==1.0.1
    # via -r requirements.in
greenlet==3.0.3
    # via
    #   eventlet
    #   sqlalchemy
itsdangerous==2.1.2
    # via
    #   flask
    #   flask-wtf
jinja2==3.1.3
    # via flask
markupsafe==2.1.5
    # via
    #   jinja2
    #   werkzeug
    #   wtforms
psutil==5.9.4
    # via -r requirements.in
psycopg2-binary==2.9.3
    # via -r requirements.in
python-dotenv==0.21.1
    # via -r requirements.in
redis==4.5.5
    # via -r requirements.in
six==1.16.0
    # via
    #   eventlet
    #   flask-cors
sqlalchemy[asyncio]==1.4.39
    # via
    #   -r requirements.in
    #   flask-sqlalchemy
    #   sqlalchemy
werkzeug==2.2.2
    # via
    #   -r requirements.in
    #   flask
    #   flask-login
wtforms==3.1.2
    # via flask-wtf
pyjwt==2.8.0
    # via -r requirements.in
